Updated various npm packages and introduced typescript
