+++ .babelrc
@@ -1,9 +1,10 @@
 {
     "presets": [
-        "env"
+        "@babel/env",
+        "@babel/preset-react"
     ],
     "plugins": [
-        "transform-object-rest-spread",
-        "transform-react-jsx"
+        "@babel/plugin-proposal-object-rest-spread",
+        "@babel/transform-react-jsx"
     ]
 }
