+++ src/interfaces/ICoordinate.ts
@@ -1,0 +1,4 @@
+export interface ICoordinate {
+  x: number;
+  y: number;
+}
