+++ webpack.config.js
@@ -1,29 +1,36 @@
-var path = require('path');
+var path = require("path");
 module.exports = {
-    entry: './src/index.js',
-    mode: 'production',
-    output: {
-        path: path.resolve(__dirname, 'build'),
-        filename: 'index.js',
-        library: 'react-image-video-lightbox',
-        libraryTarget: 'commonjs2'
-    },
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                include: path.resolve(__dirname, 'src'),
-                exclude: /(node_modules|bower_components|build)/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        presets: ['env']
-                    }
-                }
-            }
-        ]
-    },
-    externals: {
-        'react': 'commonjs react'
-    }
+  entry: "./src/index.tsx",
+  mode: "production",
+  output: {
+    path: path.resolve(__dirname, "build"),
+    filename: "index.js",
+    library: "react-image-video-lightbox",
+    libraryTarget: "commonjs2",
+  },
+  resolve: {
+    extensions: [".tsx", ".ts", ".js"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(ts|tsx)$/,
+        loader: "ts-loader",
+      },
+      {
+        test: /\.js$/,
+        include: path.resolve(__dirname, "src"),
+        exclude: /(node_modules|bower_components|build)/,
+        use: {
+          loader: "babel-loader",
+          options: {
+            presets: ["@babel/env"],
+          },
+        },
+      },
+    ],
+  },
+  externals: {
+    react: "commonjs react",
+  },
 };
